refator long

+++ src/main/java/com/integrador/dto/BairroDTO.java
@@ -13,7 +13,7 @@ import com.integrador.entities.Estado;
 public class BairroDTO implements Serializable {
 	private static final long serialVersionUID = 1L;
 	
-	private Integer id;
+	private Long id;
 	
 	@NotEmpty(message = "can't be empty")
 	@Length(min = 3, max = 80, message = "string length must be between 3 and 80")
@@ -27,7 +27,7 @@ public class BairroDTO implements Serializable {
 		
 	}
 
-	public BairroDTO(Integer id, String nome) {
+	public BairroDTO(Long id, String nome) {
 		super();
 		this.id = id;
 		this.nome = nome;
@@ -41,11 +41,11 @@ public class BairroDTO implements Serializable {
 		this.cidadeEstado = entity.getCidade().getEstado();
 	}
 	
-	public Integer getId() {
+	public Long getId() {
 		return id;
 	}
 
-	public void setId(Integer id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
